es_search_server/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-jsonfield==1.3.1
django-mptt==0.10.0
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.2
drf-extensions==0.5.0
et-xmlfile==1.0.1
jdcal==1.4.1
Jinja2==2.10.1
MarkupSafe==1.1.1
mysqlclient==1.4.4
odfpy==1.4.0
openpyxl==2.6.3
PyMySQL==0.9.3
pytz==2019.2
PyYAML==5.1.2
raven==6.10.0
redis==3.3.8
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
elasticsearch==7.0.2
